[GECO-82] added index for request ids

+++ giles-eco/src/main/java/edu/asu/diging/gilesecosystem/web/service/core/ITransactionalUploadService.java
@@ -30,7 +30,7 @@ public interface ITransactionalUploadService {
 
     int getUploadsOfUserPageCount(String username);
 
-    int getUploadsOfUserCount(String username);
+    long getUploadsOfUserCount(String username);
 
     IUpload createUpload(String username, String uploadId, String uploadDate, String uploadProgressId);
